// types/employee.ts
export interface Employee {
  id: string;
  level?: string;
  site?: string;
  type?: string;
  plantId?: string;
  createdBy?: string;
  updatedAt?: string;
  dateOfBirth?: string;
  address?: string;
  startDate?: string;
  idCard?: string;
  prefix?: string;
  firstName?: string;
  phoneNumber?: string;
  cardExpiryDate?: string;
  function?: string;
  bu?: string;
  fullName?: string;
  company?: string;
  createdAt?: string;
  empId?: string;
  displayName?: string;
  lastName?: string;
  department?: string;
  importedAt?: string;
  position?: string;
}